% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagmash.R
\name{flagmash}
\alias{flagmash}
\title{Create mash-ups of different flags
Given two ISO country codes (for country A and country B), \code{flagmash(A, B)}
will take the colours of the flag for country A and apply them to the flag
of country B. It will give the flag a name using the adjective of country A
with the name of country B.}
\usage{
flagmash(country1, country2)
}
\arguments{
\item{country1}{ISO 2 letter code of the flag to extract colours from}

\item{country2}{ISO 2 letter code of the flag to replace colours in}
}
\value{
An representation of the flag and the flag name in the RStudio viewer.
}
\description{
For example \code{flagmash("IE", "US")} will take the colours of the Republic of
Ireland's flag and apply them to the design of the flag of the United States
of America, it will call this flag "Irish United States of America".
}
